requirements.txt
pyautogui
python-whois

